Fixed the overall naming convention.
The naming convention used in this code base is called "Hungarian notation" or "CamelCase", where each word in variables and functions start with an uppercase letter.

Also, variable name's in a class starts with "m_" for stack-allocated variables and "m_p" for heap-allocated variables, indicating that the variable is a member variable of the class. This is called the "member variable" prefix.

Hungarian notation/CamelCase and member variable prefix is commonly used in game companies/studios, so make it a habit of using these.